+++ lib/dart_vm.dart
@@ -1,6 +1,8 @@
 library;
 
 export 'src/network_profile_client.dart';
+export 'src/release_updater.dart';
 export 'src/service_extension_client.dart';
 export 'src/vm_service_uri.dart';
 export 'src/vm_service_uri_config.dart';
+export 'src/vm_service_discovery.dart';